refactor: rename to prevent namespace clutter
